Allow tasks to be run with example objects.
