+++ law.cfg
@@ -33,7 +33,7 @@ wlcg_file_systems: wlcg_fs, wlcg_fs_infn_redirector, wlcg_fs_global_redirector
 
 # list of file systems used by columnflow.tasks.external.GetDatasetLFNs.iter_nano_files to
 # look for the correct fs per nano input file (in that order)
-wlcg_lfn_sources: wlcg_fs_infn_redirector, wlcg_fs_global_redirector
+wlcg_lfn_sources: wlcg_fs_desy_store, wlcg_fs_infn_redirector, wlcg_fs_global_redirector
 
 # output locations per task family
 # for local targets : "local[, STORE_PATH]"
@@ -62,6 +62,18 @@ use_cache: $CF_WLCG_USE_CACHE
 cache_root: $CF_WLCG_CACHE_ROOT
 cache_cleanup: $CF_WLCG_CACHE_CLEANUP
 cache_max_size: 50GB
+
+
+[wlcg_fs_desy_store]
+
+xrootd_base: root://dcache-cms-xrootd.desy.de:1094/pnfs/desy.de/cms/tier2
+gsiftp_base: gsiftp://dcache-door-cms04.desy.de:2811/pnfs/desy.de/cms/tier2
+base: &::xrootd_base
+use_cache: $CF_WLCG_USE_CACHE
+cache_root: $CF_WLCG_CACHE_ROOT
+cache_cleanup: $CF_WLCG_CACHE_CLEANUP
+cache_max_size: 15GB
+cache_global_lock: True
 
 
 [wlcg_fs_infn_redirector]
